Replace index page with a placeholder
The plan deploy the endorsement questionaire at a subdomain
(i.e., endorsements.inthefight.org) ran into trouble with tsl
encryption. The issue is that we need to serve the root domain
certificate from inthefight.org, and with two hosts, that's, while
certainly not impossible, non-trivial.

So instead, we mirror the main points of the current placeholder page,
so we can deploy all at once