Add the endorsement questionnaire, without ad-hoc changes (#26)
* Switch to checkboxes for partylines, update questions

* Remove todo

* WIP

* Set form field values generically

Rather than creating separate state references for each field,
share one state object, and use the event field name to set the correct
key within it.

Also add a semi-generic way to format key-value objects for Contentful.
Will stil need special cases, but since most of our fields are text, we
can share most of the work.

* WIP

* Fix field names

* Introduce basic styles to endorsements page

* Babel config

* Patch contentful-management.js

* WIP

* Fake access token

* log spaceId

* log from gatsby-node.js

* Pass tokens via siteMetadata in gatsby config

* Add intro section to endorsements page

* Use an actual ITF icon

* Remove patch for contentful-management, which mysteriously works now

* Handle reinstalls, by skipping reapplying patches

* Handle reinstalls even more, by ignoring patch's exit code

* Fix typo

* Remove Lorem Ipsum text from footer

* Bump copyright date

* Hide navigation

We're ready to launch the endorsements questionaire, but the other pages
of the site are still in dev. So hide the links leading there.

* Update test snapshots

* Unhelpful commit message

* Use correct token to send data to Contentful

* Use fonts properly and tweak weights and sizes

This grabs fonts from Adobe Typescript, via css. In a better world
we'd use `gatsby-web-fonts-loader`, and in the best of all possible
world we'd use a free alternative that we can self host. See, here
for leads regarding DIN:

https://graphicdesign.stackexchange.com/questions/7178/is-there-a-din-font-free-alternative

We probably want to do something like what fontsource does

* Fix text area label fonts, etc

More spacing fixes, and some typos

* Update test snapshots

* eslint

* Remove election date question

Note: must also be changed in Contentful. Fixes #11

* Make textArea height configurable, and use

Fixes #9.

* eslint

* Remove URL validation for campaign websites

Fixes #10

* Add numbers to questions

We do this by making the form questions a numbered list. Not so
pretty to have `<li>` inside all our form elements, and definitely
less reusable. But we'll refactor when it comes up

Fixes #6.

* Eslint, fix build, and typos

* Number sizing, and checkbox legend alignment

* Fix typo

Fixes #5.

* Switch to din-condensed (from urw-din-condensed)

* WIP: first pass at a submit alert

* Mark endorsement-questionairre fields as required

Also fix eslint issues

* Give error feedback in endorsement questionairre

* Remove debugging log msg

* Fix wrapping of site title

* Change endorsement form url

* Form numbering quick fix

We were missing a number for the District Number question. We still
need to fix the formatting, but at least with this we won't have to
renumber

* Eslint

* Add deploy script, and propagate errors from nix scripts

* Remove outdated todo

* WIP Submission message refactor

* Fix "flash of trouble" issue with submission msg

Apparently things go better if our catch error handler takes an
argument.

Fixes #12.

Co-authored-by: Ben Doyle <bdoyle@protonmail.ch>
Co-authored-by: Frida Casas <fridacasas.fc@gmail.com>